requirements.txt
fastapi
uvicorn
python-dotenv
langchain-groq